--- Assets/Scripts/OpenTS2/SimAntics/VMSleepContinueHandler.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 561701aa3ca6ba3499b29b6d198b3f80
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
